+++ feature/links/components/cards/FeaturedCard.tsx
@@ -3,7 +3,7 @@ import { Image, View } from "react-native";
 import { PressableCard } from "@/components/pressable/PressableCard";
 import { ThemedText } from "@/components/text/ThemedText";
 import { useOpenBrowser } from "@/feature/links/application/hooks/queries/browser/useOpenBrowser";
-import { type OGData } from "@/feature/links/types/links";
+import { type OGData } from "@/feature/links/domain/models/types/links";
 import { ErrorCard } from "./ErrorCard";
 
 type FeaturedArticleCardProps = {
