+++ feature/links/domain/models/index.tsx
@@ -1,0 +1,1 @@
+export * from "./types";
