+++ feature/links/domain/models/types/links.ts
